Temporary hot reloading in a more acceptable place.
